rtc: improve and move week day computing function to rtc header
Due to rtc-acpid and efi time used the same logic for computing
week day, so this patch moves code to rtc.h header file.

Additionally using a leap year algorithm to replace the for-loop
block in compute_wday for improve the performance. The first
version of algorithm is from Oliver Neukum.

Signed-off-by: Lee, Chun-Yi <jlee@suse.com>